otimizações, não foram todas
